LayoutTests/inspector-protocol/console/css-source-locations-expected.txt

Tests that CSS parser warnings from inline style tags and external stylesheets are sent to the console with correct line and column information.
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"css-source-locations.html:5:20"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"css-source-locations.html:6:14"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"css-source-locations.html:7:7"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"css-source-locations.html:9:8"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"errors.css:1:7"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"errors.css:1:29"}
{"source":"css","level":"warning","text":"Invalid CSS property declaration at: *","location":"errors.css:4:5"}
